remove resource set
